pkgs/ConfVE_Autoware/pyproject.toml
[tool.poetry]
name = "confve-autoware"
version = "0.1.0"
description = ""
authors = ["xxx"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
lark = "^1.1.9"
numpy = "^1.26.4"
shapely = "^2.0.3"
lanelet2 = "^1.2.1"
docker = "^7.0.0"
pandas = "^2.2.0"
kneed = "^0.8.5"
scikit-learn = "^1.4.1.post1"
ruamel-yaml = "^0.18.6"
ruamel-yaml-string = "^0.1.1"
networkx = "^3.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
gdown = "^5.1.0"

[tool.pytest.ini_options]
pythonpath = "src/"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
